requirements.txt
django==4.2.27
djangorestframework==3.16.1
